% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Centered log-ratio transform}
\usage{
clr(v)
}
\arguments{
\item{v}{an element or numeric vector}
}
\description{
The function returns the log-ratio transform of a vector \code{v}.
}
\details{
This function computes the log-ratio transform of a vector \code{v}. It is used by the functions that create iso-diagrams.
}
\examples{
\dontrun{clr(c(1,3,2,6))}
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data, Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). 416p.
}
\seealso{
\code{\link{clr_inv}, \link{ilr}, \link{ilr_inv}, \link{ilr3d}, \link{ilr3d_inv}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
