% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr_inv.R
\name{clr_inv}
\alias{clr_inv}
\title{Inverse centered log-ratio transform}
\usage{
clr_inv(v)
}
\arguments{
\item{v}{an element or numeric vector}
}
\description{
The function returns the inverse centered log-ratio transform of a vector \code{v}.
}
\details{
This function computes the inverse centered log-ratio transform of a vector \code{v}. It is used by the functions that create iso-diagrams.
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data, Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). 416p.
}
\seealso{
\code{\link{clr}, \link{ilr}, \link{ilr_inv}, \link{ilr3d}, \link{ilr3d_inv}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
