% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame2d.R
\name{frame2d}
\alias{frame2d}
\title{Frame for 2D ternary diagrams}
\usage{
frame2d(
  title = NULL,
  labels = c("A1", "A2", "A3"),
  lobes = TRUE,
  zones = TRUE,
  axes = TRUE
)
}
\arguments{
\item{title}{is the title of the diagram}

\item{labels}{are the labels of the three axes}

\item{lobes}{indicates whether to draw the lobes or not}

\item{zones}{indicates whether to draw the zones or not}

\item{axes}{indicates whether to draw the axes or not}
}
\description{
The function creates the frame for a 2D ternary diagram.
}
\details{
This function creates the frame for a 2D ternary diagram. It is used by the functions that create 2D ternary diagrams.
}
\seealso{
\code{\link{frame2d_iso}, \link{frame3d}, \link{frame3d_iso}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
