% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame2d_iso.R
\name{frame2d_iso}
\alias{frame2d_iso}
\title{Frame for 2D iso diagrams}
\usage{
frame2d_iso(
  title = NULL,
  labels = c("A1", "A2", "A3"),
  rad = rad,
  axes = TRUE,
  zones = TRUE,
  lobes = TRUE
)
}
\arguments{
\item{title}{is the title of the diagram}

\item{labels}{are the labels of the three axes}

\item{rad}{is a scale factor generic for the diagram}

\item{axes}{indicates whether to draw the axes or not}

\item{zones}{indicates whether to draw the zones or not}

\item{lobes}{indicates whether to draw the lobes or not}
}
\description{
The function creates the frame for a 2D iso diagram.
}
\details{
This function creates the frame for a 2D iso diagram. It is used by the functions that create 2D iso diagrams.
}
\seealso{
\code{\link{frame2d}, \link{frame3d}, \link{frame3d_iso}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
