% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame3d.R
\name{frame3d}
\alias{frame3d}
\title{Frame for 3D ternary diagrams}
\usage{
frame3d(
  title = NULL,
  labels = c("A1", "A2", "A3", "A4"),
  bg = "white",
  valpha = 0.7,
  zones = TRUE,
  lobes = TRUE,
  axes = TRUE,
  planes = FALSE
)
}
\arguments{
\item{title}{is the title of the diagram}

\item{labels}{are the labels of the four axes}

\item{bg}{is the background color}

\item{valpha}{is the longitudinal angle for the viewer}

\item{zones}{indicates whether to draw the zones or not}

\item{lobes}{indicates whether to draw the lobes or not}

\item{axes}{indicates whether to draw the axes or not}

\item{planes}{indicates whether to draw the internal planes or not}
}
\description{
The funcion creates the frame for a 3D ternary diagram.
}
\details{
This funcion creates the frame for a 3D ternary diagram. It is used by the functions that create 3D ternary diagrams.
}
\examples{
\dontrun{frame3d("3D ternary diagram")}
}
\seealso{
\code{\link{frame2d}, \link{frame2d_iso}, \link{frame3d_iso}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
