% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frame3d_iso.R
\name{frame3d_iso}
\alias{frame3d_iso}
\title{Frame for 3D iso diagrams}
\usage{
frame3d_iso(
  title = NULL,
  labels = c("A1", "A2", "A3", "A4"),
  pov = 0,
  valpha = 0.7,
  zones = TRUE,
  lobes = TRUE,
  axes = TRUE
)
}
\arguments{
\item{title}{is the title of the diagram}

\item{labels}{are the labels of the four axes}

\item{pov}{is the zenith angle for the viewer}

\item{valpha}{is the longitudinal angle for the viewer}

\item{zones}{indicates whether to draw the zones or not}

\item{lobes}{indicates whether to draw the lobes or not}

\item{axes}{indicates whether to draw the axes or not}
}
\description{
The function creates the frame for a 3D iso diagram.
}
\details{
This function creates the frame for a 3D iso diagram. It is used by the functions that create 3D iso diagrams.
}
\seealso{
\code{\link{frame2d}, \link{frame2d_iso}, \link{frame3d}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
