% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr.R
\name{ilr}
\alias{ilr}
\title{2d isometric log-ratio transform}
\usage{
ilr(v)
}
\arguments{
\item{v}{an element or numeric 3d vector}
}
\description{
The function returns the 2d isometric log-ratio transform of a vector \code{v}.
}
\details{
This function computes the 2d isometric log-ratio transform of a vector \code{v}. It is used by the functions that create iso-diagrams.
}
\examples{
\dontrun{ilr(c(1,3,0))}
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data, Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). 416p.
}
\seealso{
\code{\link{clr}, \link{clr_inv}, \link{ilr_inv}, \link{ilr3d}, \link{ilr3d_inv}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
