% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr3d.R
\name{ilr3d}
\alias{ilr3d}
\title{3d isometric log-ratio transform}
\usage{
ilr3d(v)
}
\arguments{
\item{v}{an element or numeric 4d vector}
}
\description{
The function returns the 3d isometric log-ratio transform of a vector \code{v}.
}
\details{
This function computes the 3d isomentric log-ratio transform of a vector \code{v}. It is used by the functions that create iso-diagrams.
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data, Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). 416p.
}
\seealso{
\code{\link{clr}, \link{clr_inv}, \link{ilr}, \link{ilr_inv}, \link{ilr3d_inv}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
