% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ilr_inv.R
\name{ilr_inv}
\alias{ilr_inv}
\title{Inverse 2d isometric log-ratio transform}
\usage{
ilr_inv(v)
}
\arguments{
\item{v}{an element or numeric 3d vector}
}
\description{
The function returns the inverse 2d isometric log-ratio tranform of a vector \code{v}.
}
\details{
This function computes the inverse 2d isometric log-ratio transform of a vector \code{v}. It is used by the functions that create iso-diagrams.
}
\examples{
\dontrun{ilr_inv(c(3,0,-2))}
}
\references{
Aitchison, J. (1986) The Statistical Analysis of Compositional Data, Monographs on Statistics and Applied Probability. Chapman & Hall Ltd., London (UK). 416p.
}
\seealso{
\code{\link{clr}, \link{clr_inv}, \link{ilr}, \link{ilr3d}, \link{ilr3d_inv}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
