% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso_diagram_2r.R
\name{iso_diagram_2r}
\alias{iso_diagram_2r}
\title{Isometric 2d diagram of two rounds}
\usage{
iso_diagram_2r(
  data,
  labels = NULL,
  lines = FALSE,
  color = c(2, 4, 6),
  title = "Isometric diagram",
  axes = TRUE,
  zones = TRUE,
  lobes = TRUE
)
}
\arguments{
\item{data}{is the set of preference structures of each decision maker in round 1 and round 2}

\item{labels}{is a vector with the labels of the preference structures of each decision maker}

\item{lines}{indicates whether to join the two preference structures of each decision maker by a line or not}

\item{color}{indicates the colors of symbols}

\item{title}{is the title of the diagram}

\item{axes}{indicates whether to draw the axes or not}

\item{zones}{indicates whether to draw the zones or not}

\item{lobes}{indicates whether to draw the lobes or not}
}
\description{
Isometric 2d diagram of two rounds
}
\details{
This funcion computes the isometric 2d diagram of two rounds
}
\examples{
w11 = c(0.39,0.12,0.46,0.64,0.26)
w12 = c(0.57,0.16,0.29,0.04,0.54)
w13 = c(0.04,0.73,0.25,0.32,0.20)
w21 = c(0.57,0.73,0.46,0.04,0.36)
w22 = c(0.39,0.16,0.25,0.34,0.40)
w23 = c(0.04,0.12,0.29,0.62,0.24)
dm = c('dm1','dm2','dm3','dm4','dm5')
data = data.frame(w11,w12,w13,w21,w22,w23)
iso_diagram_2r(data)
iso_diagram_2r(data,labels=dm)
iso_diagram_2r(data,lines=TRUE)
}
\references{
Turón, A., Altuzarra, A. & Moreno-Jiménez, J.M. Visual Analytics Tools for Social Cognocracy Network.
In Linden, I., Mareschal, B., Liu, S., Papathanasiou, J., & Colot, C. (Eds.). (2017). Proceedings of the 2017
International Conference on Decision Support System Technology: Data, Information and Knowledge Visualisation
in Decision Making (Vol. 3). EWG-DSS.
}
\seealso{
\code{\link{iso_diagram}, \link{iso_diagram_3d}, \link{iso_diagram_3d_2r}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
