% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_diagram.R
\name{mds_diagram}
\alias{mds_diagram}
\title{Multidimensional scaling diagram}
\usage{
mds_diagram(
  data,
  title = "MDS diagram",
  fill = TRUE,
  labels = NULL,
  color = c(1, 2, 4)
)
}
\arguments{
\item{data}{is the set of preference structures of each decision maker}

\item{title}{is the title of the diagram}

\item{fill}{indicates whether to fill the symbols or not}

\item{labels}{is a vector with the labels of the preference structures of each decision maker}

\item{color}{indicates the colors of symbols}
}
\description{
Multidimensional scaling diagram
}
\details{
This function computes the multidimensional scaling diagram (MDS) of one round
}
\examples{
w1 = c(0.29,0.24,0.15,0.15,0.27)
w2 = c(0.24,0.22,0.05,0.20,0.20)
w3 = c(0.06,0.08,0.04,0.28,0.09)
w4 = c(0.13,0.18,0.27,0.18,0.31)
w5 = c(0.28,0.28,0.49,0.20,0.12)
dm = c('dm1','dm2','dm3','dm4','dm5')
data = data.frame(w1,w2,w3,w4,w5)
mds_diagram(data)
}
\references{
Turón, A., Altuzarra, A. & Moreno-Jiménez, J.M. Visual Analytics Tools for Social Cognocracy Network.
In Linden, I., Mareschal, B., Liu, S., Papathanasiou, J., & Colot, C. (Eds.). (2017). Proceedings of the 2017
International Conference on Decision Support System Technology: Data, Information and Knowledge Visualisation
in Decision Making (Vol. 3). EWG-DSS.
}
\seealso{
\code{\link{mds_diagram_2r}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
