% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds_diagram_2r.R
\name{mds_diagram_2r}
\alias{mds_diagram_2r}
\title{Multidimensional scaling diagram of two rounds}
\usage{
mds_diagram_2r(
  data,
  title = "MDS diagram",
  fill = TRUE,
  lines = FALSE,
  labels = NULL,
  color = c(1, 2, 4, "#ff99ff")
)
}
\arguments{
\item{data}{is the set of preference structures of each decision maker in two rounds}

\item{title}{is the title of the diagram}

\item{fill}{indicates whether to fill the symbols or not}

\item{lines}{indicates whether to join the two preference structures of each decision maker by a line or not}

\item{labels}{is a vector with the labels of the preference structures of each decision maker or not}

\item{color}{indicates the colors of symbols}
}
\description{
Multidimensional scaling diagram of two rounds
}
\details{
This function computes the multidimensional scaling diagram (MDS) of two rounds
}
\examples{
w11 = c(0.29,0.24,0.15,0.15,0.27)
w12 = c(0.24,0.22,0.05,0.20,0.20)
w13 = c(0.06,0.08,0.04,0.28,0.09)
w14 = c(0.13,0.18,0.27,0.18,0.31)
w15 = c(0.28,0.28,0.49,0.20,0.12)
w21 = c(0.09,0.54,0.35,0.45,0.27)
w22 = c(0.24,0.12,0.15,0.10,0.20)
w23 = c(0.16,0.08,0.14,0.28,0.09)
w24 = c(0.13,0.18,0.17,0.08,0.31)
w25 = c(0.38,0.08,0.19,0.09,0.13)
dm = c('dm1','dm2','dm3','dm4','dm5')
data = data.frame(w11,w12,w13,w14,w15,w21,w22,w23,w24,w25)
mds_diagram_2r(data)
mds_diagram_2r(data,labels=dm)
}
\references{
Turón, A., Altuzarra, A. & Moreno-Jiménez, J.M. Visual Analytics Tools for Social Cognocracy Network.
In Linden, I., Mareschal, B., Liu, S., Papathanasiou, J., & Colot, C. (Eds.). (2017). Proceedings of the 2017
International Conference on Decision Support System Technology: Data, Information and Knowledge Visualisation
in Decision Making (Vol. 3). EWG-DSS.
}
\seealso{
\code{\link{mds_diagram}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
