% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_diagram.R
\name{ternary_diagram}
\alias{ternary_diagram}
\title{Ternary 2d diagram of one round}
\usage{
ternary_diagram(
  data,
  labels = NULL,
  color = 2,
  dens = NULL,
  title = NULL,
  lobes = TRUE,
  zones = TRUE,
  axes = TRUE
)
}
\arguments{
\item{data}{is the set of preference structures of each decision maker in one round}

\item{labels}{is a vector with the labels of the preference structures of each decision maker}

\item{color}{indicates the colors of symbols}

\item{dens}{is a vector with the weights to scale the marks by the decision makers' reputation}

\item{title}{is the title of the diagram}

\item{lobes}{indicates whether to draw the lobes or not}

\item{zones}{indicates whether to draw the zones or not}

\item{axes}{indicates whether to draw the axes or not}
}
\description{
Ternary 2d diagram of one round
}
\details{
This function computes the ternary 2d diagram of one round
}
\examples{
w1 = c(0.39,0.12,0.46,0.64,0.26)
w2 = c(0.57,0.16,0.29,0.04,0.54)
w3 = c(0.04,0.73,0.25,0.32,0.20)
dens = c(0.10,0.56,2.11,4.37,1.14)
dm = c('dm1','dm2','dm3','dm4','dm5')
data = data.frame(w1,w2,w3)
ternary_diagram(data)
ternary_diagram(data,labels=dm)
ternary_diagram(data,lobes=FALSE)
}
\references{
Turón, A., Altuzarra, A. & Moreno-Jiménez, J.M. Visual Analytics Tools for Social Cognocracy Network.
In Linden, I., Mareschal, B., Liu, S., Papathanasiou, J., & Colot, C. (Eds.). (2017). Proceedings of the 2017
International Conference on Decision Support System Technology: Data, Information and Knowledge Visualisation
in Decision Making (Vol. 3). EWG-DSS.
}
\seealso{
\code{\link{ternary_diagram_2r}, \link{ternary_diagram_3d}, \link{ternary_diagram_3d_2r}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
