% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_diagram_3d.R
\name{ternary_diagram_3d}
\alias{ternary_diagram_3d}
\title{Ternary 3d diagram of one round}
\usage{
ternary_diagram_3d(
  data,
  title = "Ternary diagram",
  labels = NULL,
  pov = 0,
  valpha = 0.7,
  zones = TRUE,
  lobes = TRUE,
  axes = TRUE,
  color = 2,
  planes = FALSE
)
}
\arguments{
\item{data}{is the set of preference structures of each decision maker in one round}

\item{title}{is the title of the diagram}

\item{labels}{is a vector with the labels of the preference structures of each decision maker}

\item{pov}{is the zenith angle for the viewer}

\item{valpha}{is the longitudinal angle for the viewer}

\item{zones}{indicates whether to draw the zones or not}

\item{lobes}{indicates whether to draw the lobes or not}

\item{axes}{indicates whether to draw the axes or not}

\item{color}{indicates the color of symbols}

\item{planes}{indicates whether to draw the internal planes or not}
}
\description{
Ternary 3d diagram of one round
}
\details{
This function computes the ternary 3d diagram of one round
}
\examples{
w1 = c(0.25,0.11,0.29,0.31,0.28)
w2 = c(0.21,0.02,0.12,0.24,0.10)
w3 = c(0.24,0.26,0.17,0.44,0.18)
w4 = c(0.31,0.61,0.42,0.01,0.44)
dm = c('dm1','dm2','dm3','dm4','dm5')
data = data.frame(w1,w2,w3,w4)
ternary_diagram_3d(data)
ternary_diagram_3d(data,labels=dm)
ternary_diagram_3d(data,lobes=FALSE)
ternary_diagram_3d(data,pov=2)
}
\references{
Turón, A., Altuzarra, A. & Moreno-Jiménez, J.M. Visual Analytics Tools for Social Cognocracy Network.
In Linden, I., Mareschal, B., Liu, S., Papathanasiou, J., & Colot, C. (Eds.). (2017). Proceedings of the 2017
International Conference on Decision Support System Technology: Data, Information and Knowledge Visualisation
in Decision Making (Vol. 3). EWG-DSS.
}
\seealso{
\code{\link{ternary_diagram}, \link{ternary_diagram_2r}, \link{ternary_diagram_3d_2r}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
