% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_diagram_3d_2r.R
\name{ternary_diagram_3d_2r}
\alias{ternary_diagram_3d_2r}
\title{Ternary 3d diagram of two rounds}
\usage{
ternary_diagram_3d_2r(
  data,
  title = NULL,
  labels = NULL,
  lines = FALSE,
  pov = 0,
  valpha = 0.7,
  dens = NULL,
  zones = TRUE,
  lobes = TRUE,
  axes = TRUE,
  color = c(2, 4, 6)
)
}
\arguments{
\item{data}{is the set of preference structures of each decision maker in round 1 and round 2}

\item{title}{is the title of the diagram}

\item{labels}{is a vector with the labels of the preference structures of each decision maker}

\item{lines}{indicates whether to join the two preference structures of each decision maker by a line or not}

\item{pov}{is the zenith angle for the viewer}

\item{valpha}{is the longitudinal angle for the viewer}

\item{dens}{indicates whether to scale the marks by the decision makers' reputation}

\item{zones}{indicates whether to draw the zones or not}

\item{lobes}{indicates whether to draw the lobes or not}

\item{axes}{indicates whether to draw the axes or not}

\item{color}{indicates the colors of symbols}
}
\description{
Ternary 3d diagram of two rounds
}
\details{
This function computes the ternary 2d diagram of two rounds
}
\examples{
w11 = c(0.25,0.11,0.29,0.31,0.28)
w12 = c(0.21,0.02,0.12,0.24,0.10)
w13 = c(0.24,0.26,0.17,0.44,0.18)
w14 = c(0.30,0.61,0.42,0.01,0.44)
w21 = c(0.21,0.12,0.42,0.61,0.08)
w22 = c(0.25,0.22,0.12,0.24,0.20)
w23 = c(0.31,0.05,0.17,0.14,0.18)
w24 = c(0.23,0.61,0.29,0.01,0.54)
dens = c(0.10,0.56,2.11,4.37,1.14)
dm = c('dm1','dm2','dm3','dm4','dm5')
data = data.frame(w11,w12,w13,w14,w21,w22,w23,w24)
ternary_diagram_3d_2r(data)
ternary_diagram_3d_2r(data,labels=dm)
ternary_diagram_3d_2r(data,lobes=FALSE)
ternary_diagram_3d_2r(data,lines=TRUE)
}
\references{
Turón, A., Altuzarra, A. & Moreno-Jiménez, J.M. Visual Analytics Tools for Social Cognocracy Network.
In Linden, I., Mareschal, B., Liu, S., Papathanasiou, J., & Colot, C. (Eds.). (2017). Proceedings of the 2017
International Conference on Decision Support System Technology: Data, Information and Knowledge Visualisation
in Decision Making (Vol. 3). EWG-DSS.
}
\seealso{
\code{\link{ternary_diagram}, \link{ternary_diagram_2r}, \link{ternary_diagram_3d}}
}
\author{
Alberto Turón \email{turon@unizar.es}
}
